Renamed Files and moved Tests Appropriately
