+++ tests/MongoDB.Analyzer.Tests/Jira/JiraTests.cs
@@ -25,8 +25,4 @@ public sealed class JiraTests : DiagnosticsTestCasesRunner
     [DataTestMethod]
     [CodeBasedTestCasesSource(typeof(VS78))]
     public Task VS78(DiagnosticTestCase testCase) => VerifyTestCase(testCase);
-
-    [DataTestMethod]
-    [CodeBasedTestCasesSource(typeof(VS114))]
-    public Task VS114(DiagnosticTestCase testCase) => VerifyTestCase(testCase);
 }
